Moved input UI to its own file instead of just being in main
